feat: BankClerk 타입 생성 및 역할 수정
